+++ core/pkg/monitor/monitor.go
@@ -4,13 +4,9 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"os"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
-
-	"golang.org/x/time/rate"
 
 	"google.golang.org/protobuf/proto"
 
@@ -22,22 +18,10 @@ import (
 )
 
 const (
-	defaultSamplingInterval = 2.0 * time.Second
-	defaultSamplesToAverage = 15
+	defaultSamplingInterval = 10.0 * time.Second
 )
 
-func Average(nums []float64) float64 {
-	if len(nums) == 0 {
-		return 0.0
-	}
-	total := 0.0
-	for _, num := range nums {
-		total += num
-	}
-	return total / float64(len(nums))
-}
-
-func makeStatsRecord(stats map[string]float64, timeStamp *timestamppb.Timestamp) *spb.Record {
+func makeStatsRecord(stats map[string]any, timeStamp *timestamppb.Timestamp) *spb.Record {
 	statsItems := make([]*spb.StatsItem, 0, len(stats))
 	for k, v := range stats {
 		jsonData, err := json.Marshal(v)
@@ -74,26 +58,9 @@ func makeMetadataRecord(metadata *spb.MetadataRequest) *spb.Record {
 	}
 }
 
-func getSlurmEnvVars() map[string]string {
-	slurmVars := make(map[string]string)
-	for _, envVar := range os.Environ() {
-		keyValPair := strings.SplitN(envVar, "=", 2)
-		key := keyValPair[0]
-		value := keyValPair[1]
-
-		if strings.HasPrefix(key, "SLURM_") {
-			suffix := strings.ToLower(strings.TrimPrefix(key, "SLURM_"))
-			slurmVars[suffix] = value
-		}
-	}
-	return slurmVars
-}
-
 type Asset interface {
 	Name() string
-	SampleMetrics() error
-	AggregateMetrics() map[string]float64
-	ClearMetrics()
+	Sample() (map[string]any, error)
 	IsAvailable() bool
 	Probe() *spb.MetadataRequest
 }
@@ -124,9 +91,6 @@ type SystemMonitor struct {
 	// The interval at which metrics are sampled
 	samplingInterval time.Duration
 
-	// The number of samples to average before sending the metrics
-	samplesToAverage int
-
 	// A logger for internal debug logging.
 	logger *observability.CoreLogger
 }
@@ -155,22 +119,17 @@ func New(
 		extraWork:        extraWork,
 		buffer:           buffer,
 		samplingInterval: defaultSamplingInterval,
-		samplesToAverage: defaultSamplesToAverage,
 	}
 
 	// TODO: rename the setting...should be SamplingIntervalSeconds
-	if si := settings.XStatsSampleRateSeconds; si != nil {
+	if si := settings.XStatsSamplingInterval; si != nil {
 		systemMonitor.samplingInterval = time.Duration(si.GetValue() * float64(time.Second))
-	}
-	if sta := settings.XStatsSamplesToAverage; sta != nil {
-		systemMonitor.samplesToAverage = int(sta.GetValue())
 	}
 
 	systemMonitor.logger.Debug(
 		fmt.Sprintf(
-			"samplingInterval: %v, samplesToAverage: %v",
+			"monitor: sampling interval: %v",
 			systemMonitor.samplingInterval,
-			systemMonitor.samplesToAverage,
 		),
 	)
 
@@ -194,6 +153,7 @@ func New(
 		NewGPUNvidia(logger, pid, samplingInterval),
 		NewGPUAMD(),
 		NewGPUApple(),
+		NewSLURM(),
 	})
 
 	return systemMonitor
@@ -218,14 +178,6 @@ func (sm *SystemMonitor) probe() *spb.MetadataRequest {
 			proto.Merge(&systemInfo, probeResponse)
 		}
 	}
-	// capture SLURM-related environment variables
-	for k, v := range getSlurmEnvVars() {
-		if systemInfo.Slurm == nil {
-			systemInfo.Slurm = make(map[string]string)
-		}
-		systemInfo.Slurm[k] = v
-	}
-
 	return &systemInfo
 }
 
@@ -311,8 +263,6 @@ func (sm *SystemMonitor) Monitor(asset Asset) {
 	ticker := time.NewTicker(sm.samplingInterval)
 	defer ticker.Stop()
 
-	sometimes := rate.Sometimes{Every: sm.samplesToAverage}
-
 	for {
 		select {
 		case <-sm.ctx.Done():
@@ -322,36 +272,33 @@ func (sm *SystemMonitor) Monitor(asset Asset) {
 				continue // Skip work when not running
 			}
 
-			// NOTE: the pattern in SampleMetric is to capture whatever metrics are available,
+			// NOTE: the pattern in Sample is to capture whatever metrics are available,
 			// accumulate errors along the way, and log them here.
-			err := asset.SampleMetrics()
+			metrics, err := asset.Sample()
 			if err != nil {
 				sm.logger.CaptureError(
 					fmt.Errorf("monitor: %v: error sampling metrics: %v", asset.Name(), err),
 				)
 			}
 
-			sometimes.Do(func() {
-				aggregatedMetrics := asset.AggregateMetrics()
-				asset.ClearMetrics()
-
-				if len(aggregatedMetrics) == 0 {
-					return // nothing to do
-				}
-				ts := timestamppb.Now()
-				// Also store aggregated metrics in the buffer if we have one
-				if sm.buffer != nil {
-					for k, v := range aggregatedMetrics {
+			if len(metrics) == 0 {
+				continue // nothing to do
+			}
+			ts := timestamppb.Now()
+			// Also store aggregated metrics in the buffer if we have one
+			if sm.buffer != nil {
+				for k, v := range metrics {
+					if v, ok := v.(float64); ok {
 						sm.buffer.push(k, ts, v)
 					}
 				}
+			}
 
-				// publish metrics
-				sm.extraWork.AddRecordOrCancel(
-					sm.ctx.Done(),
-					makeStatsRecord(aggregatedMetrics, ts),
-				)
-			})
+			// publish metrics
+			sm.extraWork.AddRecordOrCancel(
+				sm.ctx.Done(),
+				makeStatsRecord(metrics, ts),
+			)
 		}
 	}